Do not hyperlink text when LinkResolver returns null
Treat the link as broken if the LinkResolver returns null. This allows
the application developer to decide when a link is broken (which may
depend on context, as well as on $link->isBroken()), and log an error
if they like.

FakeLinkResolver has been updated to return null when
$link->isBroken() is true.

Behaviour before this patch was to render <a href="">link text</a>.

This fixes issue #21