Don't hash cache keys
This reverts most of commit 9890bb7f08c573909d1c9d4d5e8dea512a5b677c,
and its author @gsteel will need to update his Memcached
CacheInterface implementation after this patch.

There is some discussion at
https://github.com/prismicio/php-kit/issues/83

Before this patch, most cache keys were being hashed before being sent
to the CacheInterface, to limit their length.

However, though some cache backends have low key length limits, many
(including APC, which Prismic supports, and Redis and more) do not.
For these, the hashing operation is not required. It is preferable
when administrating and debugging a cache for the keys not to be
hashed, since when hashed the keys are no longer sortable in an
appreciable way and there is no clue to what the corresponding cache
values might contain.

Notably, one item cached by Prismic (the API data) did not have its
key hashed first anyway.

This patch also adds some documentation to CacheInterface to advise
developers to check if the cache backend they are writing an
implementation for has such a limit, and provides a suggestion as to
what to do if so.