Dispatch events before and after search form is submitted
A new method `Api#getDispatcher` is added, which gets an
`EventDispatcher` instance. Events can be subscribed to/listened for
via this object.

The events `prismic.pre_submit` and `prismic.post_submit` are fired
just before a search form is submitted and just after the response is
received and parsed as JSON, respectively. Both events allow the
`SearchForm` object to be retrieved, inspected and operated on, and
the `prismic.post_submit` event additionally provides the JSON
response object and a boolean denoting whether the response came from
the cache or not.

Using these events, a developer can for instance alter the form before
it is sent, or log queries and how long they take.